Revert "Initial commit"
This reverts commit b78a91ba0adf2c0716bb5818422cf117ceb2d1b9.

# Conflicts:
#	TanukiKit.xcodeproj/project.pbxproj
#	TanukiKit/User.swift
